+++ source/G3Dcpp/Matrix4.cpp
@@ -146,7 +146,7 @@ Matrix4::Matrix4(const float* init) {
 Matrix4::Matrix4(const double* init) {
     for (int r = 0; r < 4; ++r) {
         for (int c = 0; c < 4; ++c) {
-            elt[r][c] = init[r * 4 + c];
+            elt[r][c] = (float)init[r * 4 + c];
         }
     }
 }
