+++ source/G3Dcpp/Triangle.cpp
@@ -44,7 +44,7 @@ void Triangle::init(const Vector3& v0, const Vector3& v1, const Vector3& v2) {
     edge02 = _vertex[2] - _vertex[0];
 
     _primaryAxis = _plane.normal().primaryAxis();
-    _area =  edgeDirection[0].cross(edgeDirection[2]).magnitude() * (edgeMagnitude[0] * edgeMagnitude[2]);
+    _area = (float)edgeDirection[0].cross(edgeDirection[2]).magnitude() * (edgeMagnitude[0] * edgeMagnitude[2]);
 
 }
